Crate iota

IOTA core

Re-exports

pub use bee_common as common;

pub use bee_message as message;

pub use bee_pow as pow;

pub use iota_client as client;

Modules

api
High level APIs

bee_message

bee_rest_api

builder
Builder of the Client Instance

client
The Client module to connect through HORNET or Bee with API usages

error
Error handling in iota-client crate.

node
Iota node APIs

prelude

seed
The seed module

Macros

account_path
gets the BIP32 account path from a given account_index/address_internal/address_index

parse_response
match a response with an expected status code or return the default error variant.

Structs

BalanceForAddressResponse
Response of GET /api/v1/addresses/{address}

Bech32Address
Bech32 encoded address struct

Client
An instance of the client using HORNET or Bee URI

ClientBuilder
Builder to construct client instance with sensible default values

ClientMiner
The miner used for PoW

ClientMinerBuilder
The miner builder.

ConsumedOutput

CreatedOutput

Ed25519Address

Ed25519Signature

HashedIndex

IndexationPayload

Message

MessageBuilder

MessageId

MigratedFundsEntry

MilestoneIndex
A wrapper around a u32 that represents a milestone index.

MilestoneKeyRange

MilestonePayload

MilestonePayloadEssence

MilestoneResponse
Milestone data.

OutputId

OutputResponse
Response of GET /api/v1/outputs/{output_id}

ReceiptPayload

ReferenceUnlock

RegularEssence

RegularEssenceBuilder

Seed
Binary Ed25519-based Seed to derive private keys, public keys and signatures from.

SignatureLockedDustAllowanceOutput

SignatureLockedSingleOutput

TransactionId

TransactionPayload

TransactionPayloadBuilder

TreasuryInput

TreasuryOutput

TreasuryTransactionPayload

UTXOInput

Url
A parsed URL record.

Enums

Address

AddressDto

Api
Each of the node APIs the client uses.

Error
Error type of the iota client crate.

Error

Essence

Input

Output

OutputDto

Payload

SignatureUnlock

UnlockBlock

Constants

ED25519_ADDRESS_LENGTH

HASHED_INDEX_LENGTH

IOTA_SUPPLY

MESSAGE_ID_LENGTH

MESSAGE_LENGTH_MAX

MESSAGE_LENGTH_MIN

MILESTONE_MERKLE_PROOF_LENGTH

MILESTONE_PUBLIC_KEY_LENGTH

MILESTONE_SIGNATURE_LENGTH

OUTPUT_ID_LENGTH

TRANSACTION_ID_LENGTH

Functions

hash_network
Hash the network id str from the nodeinfo to an u64 for the messageBuilder

Type Definitions

Result
Type alias of Result in iota-client