Struct iota::api::ClientMessageBuilder

pub struct ClientMessageBuilder<'a> {
    client: &'a Client,
    seed: Option<&'a Seed>,
    account_index: Option<usize>,
    initial_address_index: Option<usize>,
    inputs: Option<Vec<UTXOInput>>,
    input_range: Range<usize>,
    outputs: Vec<Output>,
    index: Option<String>,
    data: Option<Vec<u8>>,
    parents: Option<Vec<MessageId>>,
}

Builder of send API

Fields
client: &'a Clientseed: Option<&'a Seed>account_index: Option<usize>initial_address_index: Option<usize>inputs: Option<Vec<UTXOInput>>input_range: Range<usize>outputs: Vec<Output>index: Option<String>data: Option<Vec<u8>>parents: Option<Vec<MessageId>>
Implementations

impl<'a> ClientMessageBuilder<'a>

pub fn new(client: &'a Client) -> ClientMessageBuilder<'a>

Create send builder

pub fn with_seed(self, seed: &'a Seed) -> ClientMessageBuilder<'a>

Sets the seed.

pub fn with_account_index(
self,
account_index: usize
) -> ClientMessageBuilder<'a>

Sets the account index.

pub fn with_initial_address_index(
self,
initial_address_index: usize
) -> ClientMessageBuilder<'a>

Sets the index of the address to start looking for balance.

pub fn with_input(self, input: UTXOInput) -> ClientMessageBuilder<'a>

Set a custom input(transaction output)

pub fn with_input_range(self, range: Range<usize>) -> ClientMessageBuilder<'a>

Set a custom range in which to search for addresses for custom inputs. Default: 0..100

pub fn with_output(
self,
address: &Bech32Address,
amount: u64
) -> Result<ClientMessageBuilder<'a>, Error>

Set a transfer to the builder, address needs to be Bech32 encoded

pub fn with_dust_allowance_output(
self,
address: &Bech32Address,
amount: u64
) -> Result<ClientMessageBuilder<'a>, Error>

Set a dust allowance transfer to the builder, address needs to be Bech32 encoded

pub fn with_output_hex(
self,
address: &str,
amount: u64
) -> Result<ClientMessageBuilder<'a>, Error>

Set a transfer to the builder, address needs to be hex encoded

pub fn with_index(self, index: &str) -> ClientMessageBuilder<'a>

Set indexation string to the builder

pub fn with_data(self, data: Vec<u8>) -> ClientMessageBuilder<'a>

Set data to the builder

pub fn with_parents(
self,
parent_ids: Vec<MessageId>
) -> Result<ClientMessageBuilder<'a>, Error>

Set 1-8 custom parent message ids

pub async fn finish(self) -> Result<Message, Error>

Consume the builder and get the API result

pub async fn finish_transaction(self) -> Result<Message, Error>

Consume the builder and get the API result

pub async fn finish_indexation(self) -> Result<Message, Error>

Consume the builder and get the API result

pub async fn finish_message(
self,
payload: Option<Payload>
) -> Result<Message, Error>

Builds the final message and posts it to the node