Module iota::bee_message::output

A module that provides types and syntactic validations of outputs.

Structs

OutputId
The identifier of an Output.

SignatureLockedDustAllowanceOutput
A SignatureLockedDustAllowanceOutput functions like a SignatureLockedSingleOutput but as a special property it is used to increase the allowance/amount of dust outputs on a given address.

SignatureLockedSingleOutput
An output type which can be unlocked via a signature. It deposits onto one single address.

TreasuryOutput
TreasuryOutput is an output which holds the treasury of a network.

Enums

Output
A generic output that can represent different types defining the deposit of funds.

Constants

DUST_THRESHOLD
Amount of tokens below which an output is considered a dust output.

OUTPUT_ID_LENGTH
The length of an OutputId.

TREASURY_OUTPUT_AMOUNT
The allowed range of the amount of a TreasuryOutput.