Module iota::bee_message::payload

A module that provides types and syntactic validations of payloads. The payload module defines the core data types for representing message payloads.

Modules

indexation
Module describing the indexation payload.

milestone
Module describing the milestone payload.

receipt
Module describing the receipt payload.

transaction
Module describing the transaction payload.

treasury
Module describing the treasury payload.

Enums

Payload
A generic payload that can represent different types defining message payloads.

Functions

option_payload_pack
Packs an optional payload to a writer.

option_payload_packed_len
Returns the packed length of an optional payload.

option_payload_unpack
Unpacks an optional payload from a reader.