Module iota::bee_message::payload::milestone

Module describing the milestone payload.

Structs

MilestoneId
A milestone identifier, the BLAKE2b-256 hash of the milestone bytes. See https://www.blake2.net/ for more information.

MilestonePayload
A payload which defines the inclusion set of other messages in the Tangle.

MilestonePayloadEssence
Essence of a milestone payload. This is the signed part of a milestone payload.

Enums

MilestoneValidationError

Constants

MILESTONE_ID_LENGTH
The length of a milestone identifier.

MILESTONE_MERKLE_PROOF_LENGTH
Length of a milestone merkle proof.

MILESTONE_PUBLIC_KEY_LENGTH
Length of a milestone public key.

MILESTONE_SIGNATURE_COUNT_RANGE
Range of allowed milestones signatures key numbers.

MILESTONE_SIGNATURE_LENGTH
Length of a milestone signature.