Module iota::bee_message::prelude

A prelude for the bee-message crate.

Structs

Ed25519Address
An Ed25519 address.

Ed25519Signature
An Ed25519 signature.

HashedIndex
Blake2b256 hash of an index.

IndexationPayload
A payload which holds an index and associated data.

Message
Represent the object that nodes gossip around the network.

MessageBuilder
A builder to build a Message.

MessageId
A message identifier, the BLAKE2b-256 hash of the message bytes. See https://www.blake2.net/ for more information.

MigratedFundsEntry
Describes funds which were migrated from a legacy network.

MilestoneId
A milestone identifier, the BLAKE2b-256 hash of the milestone bytes. See https://www.blake2.net/ for more information.

MilestoneIndex
A wrapper around a u32 that represents a milestone index.

MilestonePayload
A payload which defines the inclusion set of other messages in the Tangle.

MilestonePayloadEssence
Essence of a milestone payload. This is the signed part of a milestone payload.

OutputId
The identifier of an Output.

Parents
A [Message]'s Parents are the MessageIds of the messages it directly approves.

ReceiptPayload
Receipt is a listing of migrated funds.

ReferenceUnlock
An UnlockBlock that refers to another unlock block.

RegularEssence
A transaction regular essence consuming inputs, creating outputs and carrying an optional payload.

RegularEssenceBuilder
A builder to build a RegularEssence.

SignatureLockedDustAllowanceOutput
A SignatureLockedDustAllowanceOutput functions like a SignatureLockedSingleOutput but as a special property it is used to increase the allowance/amount of dust outputs on a given address.

SignatureLockedSingleOutput
An output type which can be unlocked via a signature. It deposits onto one single address.

TailTransactionHash
Represents a tail transaction hash of a legacy bundle.

TransactionId
A transaction identifier, the BLAKE2b-256 hash of the transaction bytes. See https://www.blake2.net/ for more information.

TransactionPayload
A transaction to move funds.

TransactionPayloadBuilder
A builder to build a TransactionPayload.

TreasuryInput
TreasuryInput is an input which references a milestone which generated a TreasuryOutput.

TreasuryOutput
TreasuryOutput is an output which holds the treasury of a network.

TreasuryTransactionPayload
TreasuryTransaction represents a transaction which moves funds from the treasury.

UnlockBlocks
A collection of unlock blocks.

UtxoInput
Represents an input referencing an output.

Enums

Address
A generic address supporting different address kinds.

Error
Error occurring when creating/parsing/validating messages.

Essence
A generic essence that can represent different types defining transaction essences.

Input
A generic input supporting different input kinds.

Output
A generic output that can represent different types defining the deposit of funds.

Payload
A generic payload that can represent different types defining message payloads.

SignatureUnlock
A SignatureUnlock contains a signature which is used to unlock a transaction input.

UnlockBlock
Defines the mechanism by which a transaction input is authorized to be consumed.

Constants

ED25519_ADDRESS_LENGTH
The number of bytes in an Ed25519 address. See https://en.wikipedia.org/wiki/EdDSA#Ed25519 for more information.

HASHED_INDEX_LENGTH
Length of an HashedIndex.

IOTA_SUPPLY
The total number of IOTA tokens in circulation.

MESSAGE_ID_LENGTH
The length of a message identifier.

MESSAGE_LENGTH_MAX
The maximum number of bytes in a message.

MESSAGE_LENGTH_MIN
The minimum number of bytes in a message.

MESSAGE_PARENTS_RANGE
The range representing the valid number of parents.

MILESTONE_ID_LENGTH
The length of a milestone identifier.

MILESTONE_MERKLE_PROOF_LENGTH
Length of a milestone merkle proof.

MILESTONE_PUBLIC_KEY_LENGTH
Length of a milestone public key.

MILESTONE_SIGNATURE_LENGTH
Length of a milestone signature.

OUTPUT_ID_LENGTH
The length of an OutputId.

TAIL_TRANSACTION_HASH_LEN
The length of a tail transaction hash.

TRANSACTION_ID_LENGTH
The length of a transaction identifier.