Module iota::bee_rest_api::types::dtos

Structs

Ed25519AddressDto
Describes an Ed25519 address.

Ed25519SignatureDto
Defines an Ed25519 signature.

GossipDto
Returns all information about the gossip stream with the peer.

HeartbeatDto
Describes the heartbeat of a node.

IndexationPayloadDto
The payload type to define a indexation payload.

MessageDto
The message object that nodes gossip around in the network.

MetricsDto
Describes metrics of a gossip stream.

MigratedFundsEntryDto

MilestonePayloadDto
The payload type to define a milestone.

PeerDto
Describes a peer.

ReceiptDto
Describes a receipt.

ReceiptPayloadDto
The payload type to define a receipt.

ReferenceUnlockDto
References a previous unlock block in order to substitute the duplication of the same unlock block data for inputs which unlock through the same data.

RegularEssenceDto
Describes the essence data making up a transaction by defining its inputs and outputs and an optional payload.

SignatureLockedDustAllowanceOutputDto
Output type for deposits that enables an address to receive dust outputs. It can be consumed as an input like a regular SigLockedSingleOutput.

SignatureLockedSingleOutputDto
Describes a deposit to a single address which is unlocked via a signature.

SignatureUnlockDto
Defines an unlock block containing signature(s) unlocking input(s).

TransactionPayloadDto
The payload type to define a value transaction.

TreasuryInputDto
Describes an input which references an unspent treasury output to consume.

TreasuryOutputDto
Describes a treasury output.

TreasuryTransactionPayloadDto
The payload type to define a treasury transaction.

UtxoInputDto
Describes an input which references an unspent transaction output to consume.

Enums

AddressDto
Describes all the different address types.

EssenceDto
Describes all the different essence types.

InputDto
Describes all the different input types.

LedgerInclusionStateDto
Describes the ledger inclusion state of a transaction.

OutputDto
Describes all the different output types.

PayloadDto
Describes all the different payload types.

RelationDto
Describes the relation with the peer.

SignatureDto
Describes all the different signature types.

UnlockBlockDto
Describes all the different unlock types.