Struct iota::client::Client

pub struct Client {
    runtime: Option<Runtime>,
    sync: Arc<RwLock<HashSet<Url, RandomState>>>,
    sync_kill_sender: Option<Arc<Sender<()>>>,
    client: Client,
    mqtt_client: Option<Client>,
    mqtt_topic_handlers: Arc<RwLock<HashMap<Topic, Vec<Arc<Box<dyn Fn(&TopicEvent) + 'static + Sync + Send, Global>>>, RandomState>>>,
    broker_options: BrokerOptions,
    network_info: Arc<RwLock<NetworkInfo>>,
    request_timeout: Duration,
    api_timeout: HashMap<Api, Duration, RandomState>,
}

An instance of the client using HORNET or Bee URI

Fields
runtime: Option<Runtime>sync: Arc<RwLock<HashSet<Url, RandomState>>>

Node pool of synced IOTA nodes

sync_kill_sender: Option<Arc<Sender<()>>>

Flag to stop the node syncing

client: Client

A reqwest Client to make Requests with

mqtt_client: Option<Client>

A MQTT client to subscribe/unsubscribe to topics.

mqtt_topic_handlers: Arc<RwLock<HashMap<Topic, Vec<Arc<Box<dyn Fn(&TopicEvent) + 'static + Sync + Send, Global>>>, RandomState>>>broker_options: BrokerOptionsnetwork_info: Arc<RwLock<NetworkInfo>>request_timeout: Duration

HTTP request timeout.

api_timeout: HashMap<Api, Duration, RandomState>

HTTP request timeout for each API call.

Implementations

impl Client

pub fn builder() -> ClientBuilder

Create the builder to instntiate the IOTA Client.

pub async fn get_network_id(&'_ self) -> Result<u64, Error>

Gets the network id of the node we're connecting to.

pub fn get_pow_provider(&self) -> ClientMiner

Gets the miner to use based on the PoW setting

pub async fn get_synced_network_info(&'_ self) -> Result<NetworkInfo, Error>

Gets the network related information and if it's the default one, sync it first.

pub async fn get_bech32_hrp(&'_ self) -> Result<String, Error>

returns the min pow score

pub async fn get_min_pow_score(&'_ self) -> Result<f64, Error>

returns the min pow score

pub fn get_tips_interval(&self) -> u64

returns the tips interval

pub fn get_local_pow(&self) -> bool

returns the local pow

pub fn get_network_info(&self) -> NetworkInfo

Gets the network related information such as network_id and min_pow_score

pub fn subscriber(&mut self) -> MqttManager<'_>

Returns a handle to the MQTT topics manager.

pub async fn get_node_health<T>(url: T) -> Result<bool, Error> where
T: IntoUrl,

GET /health endpoint

pub async fn get_health(&'_ self) -> Result<bool, Error>

GET /health endpoint

pub async fn get_node_info<T>(url: T) -> Result<InfoResponse, Error> where
T: IntoUrl,

GET /api/v1/info endpoint

pub async fn get_info(&'_ self) -> Result<InfoResponse, Error>

GET /api/v1/info endpoint

pub async fn get_peers(&'_ self) -> Result<Vec<PeerDto>, Error>

GET /api/v1/peers endpoint

pub async fn get_tips(&'_ self) -> Result<Vec<MessageId>, Error>

GET /api/v1/tips endpoint

pub async fn post_message(
&'_ self,
message: &'_ Message
) -> Result<MessageId, Error>

POST /api/v1/messages endpoint

pub fn get_message(&self) -> GetMessageBuilder<'_>

GET /api/v1/messages/{messageId} endpoint

pub async fn get_output(
&'_ self,
output_id: &'_ UTXOInput
) -> Result<OutputResponse, Error>

GET /api/v1/outputs/{outputId} endpoint Find an output by its transaction_id and corresponding output_index.

pub async fn find_outputs(
&'_ self,
outputs: &'_ [UTXOInput],
addresses: &'_ [Bech32Address]
) -> Result<Vec<OutputResponse>, Error>

Find all outputs based on the requests criteria. This method will try to query multiple nodes if the request amount exceed individual node limit.

pub fn get_address(&self) -> GetAddressBuilder<'_>

GET /api/v1/addresses/{address} endpoint

pub async fn get_milestone(
&'_ self,
index: u64
) -> Result<MilestoneResponse, Error>

GET /api/v1/milestones/{index} endpoint Get the milestone by the given index.

pub async fn get_milestone_utxo_changes(
&'_ self,
index: u64
) -> Result<MilestoneUtxoChanges, Error>

GET /api/v1/milestones/{index}/utxo-changes endpoint Get the milestone by the given index.

pub async fn reattach(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Reattaches messages for provided message id. Messages can be reattached only if they are valid and haven't been confirmed for a while.

pub async fn reattach_unchecked(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Reattach a message without checking if it should be reattached

pub async fn promote(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Promotes a message. The method should validate if a promotion is necessary through get_message. If not, the method should error out and should not allow unnecessary promotions.

pub async fn promote_unchecked(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Promote a message without checking if it should be promoted

pub fn message(&self) -> ClientMessageBuilder<'_>

A generic send function for easily sending transaction or indexation messages.

pub fn get_unspent_address(
&'a self,
seed: &'a Seed
) -> GetUnspentAddressBuilder<'a>

Return a valid unspent address.

pub fn find_addresses(&'a self, seed: &'a Seed) -> GetAddressesBuilder<'a>

Return a list of addresses from the seed regardless of their validity.

pub async fn find_messages(
&'_ self,
indexation_keys: &'_ [String],
message_ids: &'_ [MessageId]
) -> Result<Vec<Message>, Error>

Find all messages by provided message IDs and/or indexation_keys.

pub fn get_balance(&'a self, seed: &'a Seed) -> GetBalanceBuilder<'a>

Return the balance for a provided seed and its wallet chain account index. Addresses with balance must be consecutive, so this method will return once it encounters a zero balance address.

pub async fn get_address_balances(
&'_ self,
addresses: &'_ [Bech32Address]
) -> Result<Vec<BalanceForAddressResponse>, Error>

Return the balance in iota for the given addresses; No seed or security level needed to do this since we are only checking and already know the addresses.

pub async fn retry(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Retries (promotes or reattaches) a message for provided message id. Message should only be retried only if they are valid and haven't been confirmed for a while.