iota::client_ptb

Module builder

Structs

PTBBuilder

The PTBBuilder struct is the main workhorse that transforms a sequence of ParsedPTBCommands into an actual PTB that can be run. The main things to keep in mind are that this contains:

Functions

to_ordinal_contraction