iota::client_ptb

Module ptb

Structs

PTB

PTBPreview

Summary

Enums

PTBCommandResult

Functions

ptb_description

to_source_string

Convert a vector of shell tokens into a single string, with each shell token separated by a space with each command starting on a new line. NB: we add a space to the end of the source string to ensure that for unexpected EOF errors we have a location to point to.