Module iota::bee_message::prelude

Structs

ConsumedOutput

CreatedOutput

Ed25519Address

Ed25519Signature

HashedIndex

IndexationPayload

Message
A Message is the object that nodes gossip around the network.

MessageBuilder
A MessageBuilder is how you construct a Message.

MessageId
The BLAKE2b-256 hash of the byte contents of the message.

MigratedFundsEntry

MilestoneId

MilestoneIndex
A wrapper around a u32 that represents a milestone index.

MilestonePayload

MilestonePayloadEssence

OutputId

Parents
A Message's Parents are the MessageIds of the transactions it directly approves.

ReceiptPayload

ReferenceUnlock

RegularEssence

RegularEssenceBuilder

SignatureLockedDustAllowanceOutput

SignatureLockedSingleOutput

TailTransactionHash

TransactionId

TransactionPayload

TransactionPayloadBuilder

TreasuryInput

TreasuryOutput

TreasuryTransactionPayload

UnlockBlocks

UtxoInput

Enums

Address

Error

Essence

Input

Output

Payload
The payload of a message on the tangle.

SignatureUnlock

UnlockBlock

Constants

ED25519_ADDRESS_LENGTH

HASHED_INDEX_LENGTH

IOTA_SUPPLY
The total number of IOTAs in circulation.

MESSAGE_ID_LENGTH
The length of the BLAKE2b-256 hash output.

MESSAGE_LENGTH_MAX
The maximum number of bytes in a message (1024*32). https://github.com/GalRogozinski/protocol-rfcs/blob/message/text/0017-message/0017-message.md#message-validation

MESSAGE_LENGTH_MIN
The minimum number of bytes in a message.

MILESTONE_MERKLE_PROOF_LENGTH

MILESTONE_PUBLIC_KEY_LENGTH

MILESTONE_SIGNATURE_LENGTH

OUTPUT_ID_LENGTH

TAIL_TRANSACTION_HASH_LEN

TRANSACTION_ID_LENGTH