Struct iota::Client

pub struct Client {
    pub(crate) runtime: Option<Runtime>,
    pub(crate) nodes: HashSet<Url, RandomState>,
    pub(crate) sync: Arc<RwLock<HashSet<Url, RandomState>>>,
    pub(crate) sync_kill_sender: Option<Arc<Sender<()>>>,
    pub(crate) mqtt_client: Option<AsyncClient>,
    pub(crate) mqtt_topic_handlers: Arc<RwLock<HashMap<Topic, Vec<Arc<Box<dyn Fn(&TopicEvent) + 'static + Sync + Send, Global>>, Global>, RandomState>>>,
    pub(crate) broker_options: BrokerOptions,
    pub(crate) network_info: Arc<RwLock<NetworkInfo>>,
    pub(crate) request_timeout: Duration,
    pub(crate) api_timeout: HashMap<Api, Duration, RandomState>,
    pub(crate) http_client: HttpClient,
}

An instance of the client using HORNET or Bee URI

Fields
runtime: Option<Runtime>nodes: HashSet<Url, RandomState>

Node pool.

sync: Arc<RwLock<HashSet<Url, RandomState>>>

Node pool of synced IOTA nodes

sync_kill_sender: Option<Arc<Sender<()>>>

Flag to stop the node syncing

mqtt_client: Option<AsyncClient>

A MQTT client to subscribe/unsubscribe to topics.

mqtt_topic_handlers: Arc<RwLock<HashMap<Topic, Vec<Arc<Box<dyn Fn(&TopicEvent) + 'static + Sync + Send, Global>>, Global>, RandomState>>>broker_options: BrokerOptionsnetwork_info: Arc<RwLock<NetworkInfo>>request_timeout: Duration

HTTP request timeout.

api_timeout: HashMap<Api, Duration, RandomState>

HTTP request timeout for each API call.

http_client: HttpClient

HTTP client.

Implementations

impl Client

pub fn builder() -> ClientBuilder

Create the builder to instntiate the IOTA Client.

pub async fn get_network_id(&'_ self) -> Result<u64, Error>

Gets the network id of the node we're connecting to.

pub async fn get_pow_provider(&'_ self) -> ClientMiner

Gets the miner to use based on the PoW setting

pub async fn get_network_info(&'_ self) -> Result<NetworkInfo, Error>

Gets the network related information such as network_id and min_pow_score and if it's the default one, sync it first.

pub async fn get_bech32_hrp(&'_ self) -> Result<String, Error>

returns the bech32_hrp

pub async fn get_min_pow_score(&'_ self) -> Result<f64, Error>

returns the min pow score

pub async fn get_tips_interval(&'_ self) -> u64

returns the tips interval

pub async fn get_local_pow(&'_ self) -> bool

returns the local pow

pub async fn unsynced_nodes(&'_ self) -> HashSet<&'_ Url, RandomState>

returns the unsynced nodes.

pub fn subscriber(&mut self) -> MqttManager<'_>

Returns a handle to the MQTT topics manager.

pub async fn get_node_health(url: &'_ str) -> Result<bool, Error>

GET /health endpoint

pub async fn get_health(&'_ self) -> Result<bool, Error>

GET /health endpoint

pub async fn get_node_info(url: &'_ str) -> Result<InfoResponse, Error>

GET /api/v1/info endpoint

pub async fn get_info(&'_ self) -> Result<InfoResponse, Error>

GET /api/v1/info endpoint

pub async fn get_peers(&'_ self) -> Result<Vec<PeerDto, Global>, Error>

GET /api/v1/peers endpoint

pub async fn get_tips(&'_ self) -> Result<Vec<MessageId, Global>, Error>

GET /api/v1/tips endpoint

pub async fn post_message(
&'_ self,
message: &'_ Message
) -> Result<MessageId, Error>

POST /api/v1/messages endpoint

pub async fn post_message_json(
&'_ self,
message: &'_ Message
) -> Result<MessageId, Error>

POST JSON to /api/v1/messages endpoint

pub fn get_message(&self) -> GetMessageBuilder<'_>

GET /api/v1/messages/{messageId} endpoint

pub async fn get_output(
&'_ self,
output_id: &'_ UTXOInput
) -> Result<OutputResponse, Error>

GET /api/v1/outputs/{outputId} endpoint Find an output by its transaction_id and corresponding output_index.

pub async fn find_outputs(
&'_ self,
outputs: &'_ [UTXOInput],
addresses: &'_ [Bech32Address]
) -> Result<Vec<OutputResponse, Global>, Error>

Find all outputs based on the requests criteria. This method will try to query multiple nodes if the request amount exceeds individual node limit.

pub fn get_address(&self) -> GetAddressBuilder<'_>

GET /api/v1/addresses/{address} endpoint

pub async fn get_milestone(
&'_ self,
index: u32
) -> Result<MilestoneResponse, Error>

GET /api/v1/milestones/{index} endpoint Get the milestone by the given index.

pub async fn get_milestone_utxo_changes(
&'_ self,
index: u32
) -> Result<MilestoneUtxoChangesResponse, Error>

GET /api/v1/milestones/{index}/utxo-changes endpoint Get the milestone by the given index.

pub async fn get_receipts(&'_ self) -> Result<Vec<ReceiptDto, Global>, Error>

GET /api/v1/receipts endpoint Get all receipts.

pub async fn get_receipts_migrated_at(
&'_ self,
milestone_index: u32
) -> Result<Vec<ReceiptDto, Global>, Error>

GET /api/v1/receipts/{migratedAt} endpoint Get the receipts by the given milestone index.

pub async fn get_treasury(&'_ self) -> Result<TreasuryResponse, Error>

GET /api/v1/treasury endpoint Get the treasury output.

pub async fn reattach(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Reattaches messages for provided message id. Messages can be reattached only if they are valid and haven't been confirmed for a while.

pub async fn reattach_unchecked(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Reattach a message without checking if it should be reattached

pub async fn promote(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Promotes a message. The method should validate if a promotion is necessary through get_message. If not, the method should error out and should not allow unnecessary promotions.

pub async fn promote_unchecked(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Promote a message without checking if it should be promoted

pub fn message(&self) -> ClientMessageBuilder<'_>

A generic send function for easily sending transaction or indexation messages.

pub fn get_unspent_address(
&'a self,
seed: &'a Seed
) -> GetUnspentAddressBuilder<'a>

Return a valid unspent address.

pub fn get_addresses(&'a self, seed: &'a Seed) -> GetAddressesBuilder<'a>

Return a list of addresses from the seed regardless of their validity.

pub async fn find_messages<I>(
&'_ self,
indexation_keys: &'_ [I],
message_ids: &'_ [MessageId]
) -> Result<Vec<Message, Global>, Error> where
I: AsRef<[u8]>,

Find all messages by provided message IDs and/or indexation_keys.

pub fn get_balance(&'a self, seed: &'a Seed) -> GetBalanceBuilder<'a>

Return the balance for a provided seed and its wallet chain account index. Addresses with balance must be consecutive, so this method will return once it encounters a zero balance address.

pub async fn get_address_balances(
&'_ self,
addresses: &'_ [Bech32Address]
) -> Result<Vec<BalanceForAddressResponse, Global>, Error>

Return the balance in iota for the given addresses; No seed or security level needed to do this since we are only checking and already know the addresses.

pub fn parse_bech32_address(address: &str) -> Result<Address, Error>

Returns a valid Address parsed from a String.

pub fn is_address_valid(address: &str) -> bool

Checks if a String address is valid.

pub async fn retry(
&'_ self,
message_id: &'_ MessageId
) -> Result<(MessageId, Message), Error>

Retries (promotes or reattaches) a message for provided message id. Message should only be retried only if they are valid and haven't been confirmed for a while.

pub async fn retry_until_included(
&'_ self,
message_id: &'_ MessageId,
interval: Option<u64>,
max_attempts: Option<u64>
) -> Result<Vec<(MessageId, Message), Global>, Error>

Retries (promotes or reattaches) a message for provided message id until it's included (referenced by a milestone). Default interval is 5 seconds and max attempts is 10. Returns reattached messages